framework/wifi_manager : Fix build error due to vwifi_start undefined symbol
The vwifi_start symbol is defined in kernel. So, when loadable_apps config
is used, this symbol cannot be accessed directly from app binary, and needs
to be accessed via syscall or ioctl. As a temporary fix, we check for
CONFIG_BUILD_FLAT in the framework code to avoid linker error.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>